requirements.txt
flask
flask_cors
requests
bs4
lxml
pickle
sklearn
mysql-connector-python
gunicorn
